Separate integration tests from unit tests
