+++ .travis.yml
@@ -14,7 +14,6 @@ cache:
   directories:
   - "~/.ivy2"
   apt: true
-sbt_args: "-Dsbt.log.noformat=true -sbt-dir ~/.sbt"
 after_success:
 - sh "$TRAVIS_BUILD_DIR/.travis.release.sh"
 env:
@@ -23,3 +22,5 @@ env:
   - secure: w7Ck9Si1upg4pX3h4r7PzUUe9lVegPISLzOaey2WEi6GVZfyx0jLKVJf+ePmLG+TajNxMU8nFwhUdnJNl70NABzvxd0enywYwfZ6XnnntVzEQBrXhXVymTxv20CD5JlCUx28k2m6AVTEaizIFb8YpeHs2gxTh+EDW2B8bg/sSyqvloSoNjLbpYfdE2MYc1P6k1IR/4jG8ofyHxNxiJFtPEz3xMra/CnuH1p2l5Sx0IHdB3YlrLoHQV/hwePhTzAsPCb/kHputfV+LfKX6bL6zHC1d5lmgKUWbfn96jKzoSCAG5Oo1ulJqQtLCNDasMre2IzDcQBX41YObGdr6HlBfRpVmKKOt5Ui2h8JO/Sl4Fc+O2VhuEoa3uwXF4L7LnlPRY/kQJ8U3AM2DIFjHcN8LS4Inz6Lrsfe133cN+kUruQoBQOuFTNjkTPH3F2W/H4ZVf+rfKc2mJ+SubrYefXknwYwlXKTn76WSCV0Dwm9ff0aqrRGPEpEk1ym0UOvYEmQSKcpU6jDZYwIBWsa5qt9TzSjbyxtQMS1CqplFr+3Dmw5xZNrFR/VlimBGFWwJf1wGwwZN8FZ60pF6mkmtBkVRqc8KQtqkkkXTVpjeBhHE+XubwqCnxTlM0vr0TzhqBTosVb5NeOscQYnvNOyY833uV2zyoQae+ZRN0SqHaZT3xQ=
   - secure: KeHDYk6ExVwpKz7GZdTJg8xiUTGqBRxOOWvPE/d7UQUNRHfs7YbcWsWx3oqhoLvMPb5KbXr27q25B/vIrmkES5RBZXYTr3AlmlDPslnUVsoxsG1oqH79lBk5z2lnZCBTB1BgWq2cOXLpavXHoPGPMzI9KKJqr3zgWnruDGWzgzAJx0X2UhAsL6VLe7QbLHRM3JALpbZ0y+6DDRxtklwNAePDcwaMbm6eCPN7+lbXhIMee0pbqyRex6JWOukReAaqA/cGM4tnxOvIfN9jhH5bcYI0ufzH8oKa0s85aP3EFZsYZjSDn2amvKMDWhGNQENPaMU51nbQZTvWccdA+KqmYc/SVTfJvmIDTIFBDNMtdqfLdKLaVBtGLZNktKfJRW4vXbjfWtBUuoMXWWwS4w7XvKuX1mM5zh7R2OPN7j7L6uP3VExVf3uojGJvnslYwGIDqMyUSW8uhEOe0fQIm/9dozxqYBuHSZeZ9sEyn2wAWQZR84B4x5PrqvhfPQwhL0TPrE5Y1InxnE63UI5Dc11nIwImhUu+Q7aw0IDAWsyTGi2FsFvQ+f2tPB/ypMbA//G+DcPwFNJqxcj2fXJfDRLvgS0rUozkAU1BGJjzCfkRJ8FHIs6RSUfNBHBzEr/sb9m9nUqDzWvV1bao2FlU1LAc0NjfkcoBamsce4DhieywWGs=
   - secure: ZJdC/n659AbfZ98Hecv7WuJd2/E+BNZUIghsKG96cGtK48FMYc+6YBfV+b50I4M8bS62nq+rqeIwN6QLEecQHASsJGzWK2T6fpqubbPbZO0zo2AfC+4wzjDnh1ytfGJLLbmYREuQjRwVBJRa3A9yKMCpmPyTEOVwCIwkbEIgO4ngHWuLtvMQm4lA/YjeBz54+APaLHmEvCj3eJ9nidVc0+9M+7iwuEbACr+jyTqb145y/ZEhUbLC2TTruIwLYaHEroOc9tQyV56MYjd95VJ5gLNg5+FVqcQe2dmiDswwOKn1GOv2/BJjfL4sDXDbyde7kmJcghEfmSjJ38E1lcAHBuHYT/lU50gk3Uc0U6qAkzo0mpwhrEOlfD8fLOhIh3Qu8rDy6U3LDvVR8ckwq0Nb6VYbyGwJJiwCxnDxkMe2vut04fgy4FzVTCNBcnuCa0zgCUTgoUy+EHcVoYlfHzjrRH3zft0NSsa3S94hFFFfhXFjwN1Bz6/eBCIRk6Brk4/z0lD69k9e+QlsmA6DngZuCvSzYB8ZiD89LyabNSojZWSdYS1i9jmnkYUpDBu6hrU5FtGAXyyYiRi8kGUSwNm6jmFGQ0QiijR9ZVzJdxwl7oUUebU67aPk/KTmUBItuCikxU78JZDmYGEHpfMdb/qZDRmnf4ARprsSagGgtxhdRIM=
+script:
+  - sbt -Dsbt.log.noformat=true -sbt-dir ~/.sbt ++$TRAVIS_SCALA_VERSION test it:test
