+++ build.sbt
@@ -8,19 +8,26 @@ resolvers += "bintray-pagerduty-oss-maven" at "https://dl.bintray.com/pagerduty/
 
 // Prevents logging configuration from being included in the test jar.
 mappings in (Test, packageBin) ~= { _.filterNot(_._2.endsWith("logback-test.xml")) }
+mappings in (IntegrationTest, packageBin) ~= { _.filterNot(_._2.endsWith("logback-it.xml")) }
 
 // Dependencies in this configuration are not exported.
 ivyConfigurations += config("transient").hide
 
 fullClasspath in Test ++= update.value.select(configurationFilter("transient"))
 
-libraryDependencies ++= Seq(
-  "com.pagerduty" %% "eris-core" % "1.5.1",
-  "com.pagerduty" %% "entity-mapper" % "0.6.1")
+lazy val root = (project in file(".")).
+  configs(IntegrationTest extend (Test)).
+  settings(Defaults.itSettings: _*).
+  settings(
 
-libraryDependencies ++= Seq(
-  "ch.qos.logback" % "logback-classic" % "1.0.13" % "transient",
-  "com.pagerduty" %% "eris-core" % "1.5.1" % Test classifier "tests",
-  "org.scalatest" %% "scalatest" % "2.2.4" % Test,
-  "org.scalamock" %% "scalamock-scalatest-support" % "3.2" % Test,
-  "org.scalacheck" %% "scalacheck" % "1.12.2" % Test)
+  libraryDependencies ++= Seq(
+    "com.pagerduty" %% "eris-core" % "1.5.1",
+    "com.pagerduty" %% "entity-mapper" % "0.6.1"),
+
+  libraryDependencies ++= Seq(
+    "ch.qos.logback" % "logback-classic" % "1.0.13" % "transient",
+    "com.pagerduty" %% "eris-core" % "1.5.1" % Test classifier "tests",
+    "org.scalatest" %% "scalatest" % "2.2.4" % "it,test",
+    "org.scalamock" %% "scalamock-scalatest-support" % "3.2" % "it,test",
+    "org.scalacheck" %% "scalacheck" % "1.12.2" % "it,test")
+  )
