+++ version.sbt
@@ -1,1 +1,1 @@
-version := "1.6.1"
+version := "1.6.2"
